Add games count functionality
Using the very advanced SQLite function Count I am beginning to add the
total count of games recorded since the inception of the game. As a PM
there is nothing more important to me than understanding and tracking
where all my time goes and I think total ping pong games is a good
approximation.

This is a multi part change that will require other changes potentially
on the matches page as well and may include styling or position updates
after consulting with the ping pong PM.

In the future we might be able to add more advanced functionality like a
table of who plays the most games but will also need to consult the ping
pong PM to see if it's on the backlog.